data-adapter/requirements.txt
pandas==1.3.5
paho-mqtt==1.6.1